# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Website Discovery & Knowledge Graph Builder",
    "description": "Crawls a target website, discovers pages/navigation/UI elements, and builds a relational + graph knowledge base (Postgres + Neo4j) describing the site's structure and workflows.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "tags": [
    { "name": "Health", "description": "Service health check" },
    { "name": "Auth", "description": "User authentication, Google/GitHub OAuth sync, and audit logs" },
    { "name": "Crawl", "description": "Queue and track website crawl jobs" },
    { "name": "Knowledge", "description": "Read back the extracted knowledge graph and summary" }
  ],
  "paths": {
    "/api/auth/sync": {
      "post": {
        "tags": ["Auth"],
        "summary": "Synchronize user profile from Google or GitHub OAuth",
        "description": "Upserts user in PostgreSQL and automatically records a SIGNUP or LOGIN audit event.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "provider"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@gmail.com" },
                  "name": { "type": "string", "example": "Alex Mercer" },
                  "avatarUrl": { "type": "string", "example": "https://lh3.googleusercontent.com/..." },
                  "provider": { "type": "string", "enum": ["google", "github", "demo"], "example": "google" },
                  "providerUserId": { "type": "string", "example": "google_sub_12345" },
                  "role": { "type": "string", "example": "user" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated and login logged successfully" },
          "201": { "description": "User registered and signup logged successfully" }
        }
      }
    },
    "/api/auth/users": {
      "get": {
        "tags": ["Auth"],
        "summary": "List registered users",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of users with last login status" }
        }
      }
    },
    "/api/auth/logs": {
      "get": {
        "tags": ["Auth"],
        "summary": "List authentication audit logs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Recent signups, logins, and logouts" }
        }
      }
    },
    "/api/auth/stats": {
      "get": {
        "tags": ["Auth"],
        "summary": "Aggregated user and authentication statistics",
        "responses": {
          "200": { "description": "Total users, signups, logins, and provider breakdown" }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "healthy" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/crawl": {
      "post": {
        "tags": ["Crawl"],
        "summary": "Queue a website discovery crawl",
        "description": "Inserts a PENDING crawl_jobs row. The crawl-worker service polls for pending jobs, runs the Playwright crawl, and builds the knowledge graph asynchronously. If this exact targetUrl already has a COMPLETED crawl within CRAWL_TTL_HOURS (default 24h), that existing job/project is returned instead (response.cached=true) rather than queuing a new crawl.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targetUrl"],
                "properties": {
                  "targetUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com"
                  },
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Optional. Reuse an existing project ID to re-crawl into the same project; omit to start a new project."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crawl job queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CrawlJobResponse" }
              }
            }
          },
          "200": {
            "description": "An existing COMPLETED crawl for this targetUrl is still within CRAWL_TTL_HOURS; it was returned instead of queuing a new crawl (response.cached=true)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CrawlJobResponse" }
              }
            }
          },
          "400": { "description": "targetUrl is required" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/crawl/{id}": {
      "get": {
        "tags": ["Crawl"],
        "summary": "Get crawl job status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Crawl job ID returned by POST /api/crawl"
          }
        ],
        "responses": {
          "200": {
            "description": "Crawl job status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CrawlJob" }
              }
            }
          },
          "404": { "description": "Crawl job not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/crawl/{id}/credentials": {
      "post": {
        "tags": ["Crawl"],
        "summary": "Submit login credentials for a paused crawl",
        "description": "When a crawl hits a login wall on its start page, the job pauses with status AWAITING_CREDENTIALS and login_url set. Submit credentials here to resume it -- the worker logs in with Playwright and continues the crawl authenticated. Credentials are held only transiently: the worker deletes them from the database as soon as it reads them, whether or not the login attempt succeeds.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Crawl job ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "demo@example.com" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Credentials accepted; job set back to PENDING for the worker to retry" },
          "400": { "description": "username and password are required" },
          "404": { "description": "Crawl job not found" },
          "409": { "description": "Job is not currently awaiting credentials" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/graph/{projectId}": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "Get the project's knowledge graph",
        "description": "Returns all Neo4j nodes (Page, Entity, Action, Workflow, WorkflowStep) and edges (HAS_PAGE/CHILD_OF with the clicked label+selector, RELATED_TO, HAS_ACTION, HAS_STEP, NEXT_STEP, etc.) for the project.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Graph projection",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GraphResponse" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/summary/{projectId}": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "Get the project's high-level knowledge summary",
        "description": "Returns a lightweight domain summary (main entities, main workflows, discovered page count) intended to save downstream AI token usage compared to reading the full graph.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SummaryResponse" }
              }
            }
          },
          "404": { "description": "Summary not found for this project" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CrawlJob": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "project_id": { "type": "string", "format": "uuid" },
          "target_url": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "RUNNING", "AWAITING_CREDENTIALS", "ENRICHING", "COMPLETED", "FAILED"]
          },
          "login_url": {
            "type": "string",
            "nullable": true,
            "description": "Set when status is AWAITING_CREDENTIALS: the page that needs login. POST credentials to /api/crawl/{id}/credentials to resume."
          },
          "started_at": { "type": "string", "format": "date-time", "nullable": true },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true },
          "error_message": { "type": "string", "nullable": true }
        }
      },
      "CrawlJobResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Crawl job queued successfully" },
          "cached": { "type": "boolean", "description": "True if an existing completed crawl within CRAWL_TTL_HOURS was returned instead of queuing a new one" },
          "job": { "$ref": "#/components/schemas/CrawlJob" }
        }
      },
      "GraphNode": {
        "type": "object",
        "properties": {
          "labels": { "type": "array", "items": { "type": "string" } },
          "properties": { "type": "object" }
        }
      },
      "GraphEdge": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "example": "HAS_PAGE" },
          "source": { "type": "string" },
          "target": { "type": "string" },
          "properties": { "type": "object" }
        }
      },
      "GraphResponse": {
        "type": "object",
        "properties": {
          "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/GraphNode" } },
          "edges": { "type": "array", "items": { "$ref": "#/components/schemas/GraphEdge" } }
        }
      },
      "SummaryResponse": {
        "type": "object",
        "properties": {
          "domain": { "type": "string" },
          "summary_data": {
            "type": "object",
            "properties": {
              "domain": { "type": "string" },
              "timestamp": { "type": "string", "format": "date-time" },
              "mainEntities": { "type": "array", "items": { "type": "string" } },
              "mainWorkflows": { "type": "array", "items": { "type": "object" } },
              "discoveredPagesCount": { "type": "integer" }
            }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}